styling(frontend): Renames files to use kebab-case for better readability and makes the dropdown-menu-item a part of the dropdown-menu. UT-188
