--- frontend/src/admin/actions.json
@@ -1,16 +1,0 @@
-{
-  "actions": [
-    {
-      "id": "1",
-      "name": "Download user"
-    },
-    {
-      "id": "2",
-      "name": "Upload user"
-    },
-    {
-      "id": "3",
-      "name": "Add user"
-    }
-  ]
-}
