+++ frontend/src/admin/components/sidebar/dropdown/dropdown-menu-item.scss
@@ -1,0 +1,19 @@
+@import "src/styles/main";
+
+p {
+  color: $contrast-color;
+  height: 30px;
+  display: flex;
+  align-items: center;
+  padding: 1rem;
+  border-bottom: 1px solid $border-color;
+
+  &:last-child {
+    border-bottom: none;
+  }
+
+  &:hover {
+    font-weight: bold;
+    transition:  0.3s ease-in-out;
+  }
+}
