--- frontend/src/admin/components/sidebar/dropdown/dropdown.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-import Actions from "../../../actions.json";
-import DropdownItem from "./dropdownItem";
-
-export default function Dropdown() {
-    return (
-        <div className="c-dropdown">
-            {Actions.actions.map((action) => (
-                <DropdownItem
-                    action={action}
-                    key={action.id}
-                />
-            ))}
-        </div>
-    );
-}
