--- frontend/src/admin/components/sidebar/dropdown/dropdownItem.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import {Action} from "../../../models/action";
-
-interface DropDownProps {
-    action: Action
-}
-
-export default function DropdownItem(props: DropDownProps) {
-    return (
-        <p>
-            {props.action.name}
-        </p>
-    )
-}
