+++ frontend/src/admin/components/sidebar/dropdown/menu-icon.tsx
