--- frontend/src/admin/models/action.ts
@@ -1,4 +1,0 @@
-export interface Action {
-    id: string,
-    name: string
-}
